+++ Editor/Utilities/BuildCache.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Threading;
 using UnityEditor.Build.Content;
@@ -17,6 +18,7 @@ namespace UnityEditor.Build.Pipeline.Utilities
     public class BuildCache : IBuildCache, IDisposable
     {
         const string k_CachePath = "Library/BuildCache";
+        const int k_Version = 2;
 
         Dictionary<KeyValuePair<GUID, int>, CacheEntry> m_GuidToHash = new Dictionary<KeyValuePair<GUID, int>, CacheEntry>();
         Dictionary<KeyValuePair<string, int>, CacheEntry> m_PathToHash = new Dictionary<KeyValuePair<string, int>, CacheEntry>();
@@ -60,7 +62,7 @@ namespace UnityEditor.Build.Pipeline.Utilities
 
         static Hash128 CalculateGlobalArtifactVersionHash()
         {
-            return HashingMethods.Calculate(PlayerSettings.scriptingRuntimeVersion, Application.unityVersion).ToHash128();
+            return HashingMethods.Calculate(PlayerSettings.scriptingRuntimeVersion, Application.unityVersion, k_Version).ToHash128();
         }
 
         internal void ClearCacheEntryMaps()
@@ -97,11 +99,12 @@ namespace UnityEditor.Build.Pipeline.Utilities
             else
             {
                 entry.Hash = AssetDatabase.GetAssetDependencyHash(path);
-                if (!entry.Hash.isValid)
+                if (!entry.Hash.isValid && File.Exists(path))
                     entry.Hash = HashingMethods.CalculateFile(path).ToHash128();
             }
 
-            entry.Hash = HashingMethods.Calculate(entry.Hash, entry.Version).ToHash128();
+            if (entry.Hash.isValid)
+                entry.Hash = HashingMethods.Calculate(entry.Hash, entry.Version).ToHash128();
 
             m_GuidToHash[key] = entry;
             return entry;
@@ -115,9 +118,14 @@ namespace UnityEditor.Build.Pipeline.Utilities
             if (m_PathToHash.TryGetValue(key, out entry))
                 return entry;
 
+            var guid = AssetDatabase.AssetPathToGUID(path);
+            if (!string.IsNullOrEmpty(guid))
+                return GetCacheEntry(new GUID(guid), version);
+
             entry = new CacheEntry { File = path, Version = version };
             entry.Guid = HashingMethods.Calculate("FileHash", entry.File).ToGUID();
-            entry.Hash = HashingMethods.Calculate(HashingMethods.CalculateFile(entry.File), entry.Version).ToHash128();
+            if (File.Exists(entry.File))
+                entry.Hash = HashingMethods.Calculate(HashingMethods.CalculateFile(entry.File), entry.Version).ToHash128();
             entry.Type = CacheEntry.EntryType.File;
 
             m_PathToHash[key] = entry;
@@ -161,7 +169,7 @@ namespace UnityEditor.Build.Pipeline.Utilities
         {
             var guid = entry.Guid.ToString();
             string finalHash = HashingMethods.Calculate(m_GlobalHash, entry.Hash).ToString();
-            return string.Format("{0}/{1}/{2}/{2}_{3}.info", k_CachePath, guid.Substring(0, 2), guid, finalHash);
+            return string.Format("{0}/{1}/{2}/{3}/{2}.info", k_CachePath, guid.Substring(0, 2), guid, finalHash);
         }
 
         /// <inheritdoc />
@@ -169,7 +177,7 @@ namespace UnityEditor.Build.Pipeline.Utilities
         {
             var guid = entry.Guid.ToString();
             string finalHash = HashingMethods.Calculate(m_GlobalHash, entry.Hash).ToString();
-            return string.Format("{0}/{1}/{2}", k_CachePath, guid.Substring(0, 2), finalHash);
+            return string.Format("{0}/{1}/{2}/{3}", k_CachePath, guid.Substring(0, 2), guid, finalHash);
         }
 
         class FileOperations
@@ -353,26 +361,116 @@ namespace UnityEditor.Build.Pipeline.Utilities
 
         internal void SyncPendingSaves()
         {
-            if(m_ActiveWriteThread != null)
+            if (m_ActiveWriteThread != null)
             {
                 m_ActiveWriteThread.Join();
                 m_ActiveWriteThread = null;
             }
         }
 
-        [MenuItem("Window/Asset Management/Purge Build Cache", priority = 10)]
-        public static void PurgeCache()
+        internal struct CacheFolder
         {
-            PurgeCache(false);
+            public DirectoryInfo directory;
+            public long Length { get; set; }
+            public void Delete() => directory.Delete(true);
+            public DateTime LastAccessTimeUtc
+            {
+                get => directory.LastAccessTimeUtc;
+                internal set => directory.LastAccessTimeUtc = value;
+            }
         }
 
         public static void PurgeCache(bool prompt)
         {
-            if (prompt && !EditorUtility.DisplayDialog("Purge Build Cache", "Do you really want to purge your entire build cache?", "Yes", "No"))
+            if (!Directory.Exists(k_CachePath))
+            {
+                if (prompt)
+                    Debug.Log("Current build cache is empty.");
+                return;
+            }
+
+            if (prompt)
+            {
+                if (!EditorUtility.DisplayDialog("Purge Build Cache", "Do you really want to purge your entire build cache?", "Yes", "No"))
+                    return;
+
+                EditorUtility.DisplayProgressBar(BuildCachePreferences.BuildCacheProperties.purgeCache.text, BuildCachePreferences.BuildCacheProperties.pleaseWait.text, 0.0F);
+                Directory.Delete(k_CachePath, true);
+                EditorUtility.ClearProgressBar();
+            }
+            else
+                Directory.Delete(k_CachePath, true);
+        }
+
+        public static void PruneCache()
+        {
+            int maximumSize = EditorPrefs.GetInt("BuildCache.maximumSize", 200);
+            long maximumCacheSize = maximumSize * 1073741824L; // gigabytes to bytes
+
+            // Get sizes based on common directory root for a guid / hash
+            ComputeCacheSizeAndFolders(out long currentCacheSize, out List<CacheFolder> cacheFolders);
+
+            if (currentCacheSize < maximumCacheSize)
+            {
+                Debug.LogFormat("Current build cache currentCacheSize {0}, prune threshold {1} GB. No prune performed. You can change this value in the \"Edit/Preferences...\" window.", EditorUtility.FormatBytes(currentCacheSize), maximumSize);
+                return;
+            }
+
+            if (!EditorUtility.DisplayDialog("Prune Build Cache", string.Format("Current build cache currentCacheSize is {0}, which is over the prune threshold of {1}. Do you want to prune your build cache now?", EditorUtility.FormatBytes(currentCacheSize), EditorUtility.FormatBytes(maximumCacheSize)), "Yes", "No"))
                 return;
 
-            if (Directory.Exists(k_CachePath))
-                Directory.Delete(k_CachePath, true);
+            EditorUtility.DisplayProgressBar(BuildCachePreferences.BuildCacheProperties.pruneCache.text, BuildCachePreferences.BuildCacheProperties.pleaseWait.text, 0.0F);
+
+            PruneCacheFolders(maximumCacheSize, currentCacheSize, cacheFolders);
+
+            EditorUtility.ClearProgressBar();
+        }
+
+        public static void PruneCache_Background(long maximumCacheSize)
+        {
+            // Get sizes based on common directory root for a guid / hash
+            ComputeCacheSizeAndFolders(out long currentCacheSize, out List<CacheFolder> cacheFolders);
+            if (currentCacheSize < maximumCacheSize)
+                return;
+
+            PruneCacheFolders(maximumCacheSize, currentCacheSize, cacheFolders);
+        }
+
+        internal static void ComputeCacheSizeAndFolders(out long currentCacheSize, out List<CacheFolder> cacheFolders)
+        {
+            currentCacheSize = 0;
+            cacheFolders = new List<CacheFolder>();
+
+            var directory = new DirectoryInfo(k_CachePath);
+            if (!directory.Exists)
+                return;
+
+            int length = directory.FullName.Count(x => x == Path.DirectorySeparatorChar) + 3;
+            DirectoryInfo[] subDirectories = directory.GetDirectories("*", SearchOption.AllDirectories);
+            foreach (var subDirectory in subDirectories)
+            {
+                if (subDirectory.FullName.Count(x => x == Path.DirectorySeparatorChar) != length)
+                    continue;
+
+                FileInfo[] files = subDirectory.GetFiles("*", SearchOption.AllDirectories);
+                var cacheFolder = new CacheFolder { directory = subDirectory, Length = files.Sum(x => x.Length) };
+                cacheFolders.Add(cacheFolder);
+
+                currentCacheSize += cacheFolder.Length;
+            }
+        }
+
+        internal static void PruneCacheFolders(long maximumCacheSize, long currentCacheSize, List<CacheFolder> cacheFolders)
+        {
+            cacheFolders.Sort((a, b) => a.LastAccessTimeUtc.CompareTo(b.LastAccessTimeUtc));
+            // Need to delete sets of files as the .info might reference a specific file artifact
+            foreach (var cacheFolder in cacheFolders)
+            {
+                currentCacheSize -= cacheFolder.Length;
+                cacheFolder.Delete();
+                if (currentCacheSize < maximumCacheSize)
+                    break;
+            }
         }
     }
 }