+++ Tests/Editor/CacheServerBuildCacheTests.cs
@@ -21,9 +21,9 @@ namespace UnityEditor.Build.Pipeline.Tests
         {
             // purge the local cache to make sure we don't load anything out of it again.
             // these tests need to pull from the cache server
-            BuildCache.PurgeCache();
             if (m_Cache != null)
                 m_Cache.Dispose();
+            PurgeBuildCache();
             m_Cache = new BuildCache("localhost", LocalCacheServer.instance.m_port);
         }
     }