+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1-preview] - 2019-04-16
+- Fixed "Path is empty" exception in build cache
+- Fixed an edge case where a valid cache entry could be returned for an invalid request
+- Added BuildCache.PruneCache API to trim the cache down to a limit, called in the background after a build
+- Moved BuildCache menu options and preferences to the "Edit/Prefferences..." window
+
 ## [1.3.5-preview] - 2019-02-28
 - Minimum Unity version is now 2018.3 to address a build-time bug with progressive lightmapper.
 - Added missing version into CacheEntry calculation